+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 
 # CHANGELOG
 
+## 2 : optimized makeKey()
+- optimized the code for `makeKey()`
+
 ## 1 : initial commit
 - initial commit